Download values at given coordinates

Downloads data for a point (coordinate) in a grid cell (coordinate system: (EPSG: 25833)) at the given time resolution. Several coordinates can be provided at once. Time resolution is given by the choice of environmental variable (env_layer) and start_date/end_date. E.g. "tm3h" for "2023-12-01T06" is three hour data for temperature collected at 06:00 and "tm" for "2023-12-01" is the daily average temperature.

Note 1: The API has a cap at ~635'000 values in each download and ends with an error for larger queries.

Note 2: The API ends with an error if some of the points have the same cellindex. Use gts_cellindex() to check.

Usage

gts_dl_coords(
  coords,
  env_layer,
  start_date,
  end_date,
  return_raw = FALSE,
  verbose = FALSE
)

Arguments

coords

A data frame with x and y coordinates (coordinate system: (EPSG: 25833))

env_layer

The quoted abbreviation for the environmental layer to download. E.g. Daily precipitaion = "rr", Temperature = "tm", Snow depth = "sd".

start_date

The start date given as: 'YYYY-MM-DD'. If querying three or one hour data, the hour should be given like this: format: 'YYYY-MM-DDT06'

end_date

The end date. See start_date.

return_raw

Set to TRUE to return the raw results from the query (a list) without coercing to a data.frame.

verbose

Set to TRUE to print the query to the console (default FALSE).

Value

A data frame with the GTS data associated with each point for each step in the time resolution, e.g. daily values or values for each hour.

Examples

## Create a data frame with coordinates
coords <- data.frame(x = c(58200, 74225), y = c(6708225, 6715025))
## Download data
gts_dl_coords(coords = coords, env_layer = "tm", start_date = "2023-12-01", end_date = "2023-12-02")
#>       x       y cellindex altitude       date  time time_resolution_minutes
#> 1 58200 6708225   1542878     1294 2023-12-01 00:00                    1440
#> 2 58200 6708225   1542878     1294 2023-12-02 00:00                    1440
#> 3 74225 6715025   1534529     1371 2023-12-01 00:00                    1440
#> 4 74225 6715025   1534529     1371 2023-12-02 00:00                    1440
#>      unit    tm
#> 1 Celcius -13.8
#> 2 Celcius -13.0
#> 3 Celcius -15.2
#> 4 Celcius -14.1
gts_dl_coords(coords = coords, env_layer = "tm3h", start_date = "2023-12-01T23", end_date = "2023-12-02T06")
#>       x       y cellindex altitude       date  time time_resolution_minutes
#> 1 58200 6708225   1542878     1294 2023-12-01 23:00                     180
#> 2 58200 6708225   1542878     1294 2023-12-02 02:00                     180
#> 3 58200 6708225   1542878     1294 2023-12-02 05:00                     180
#> 4 74225 6715025   1534529     1371 2023-12-01 23:00                     180
#> 5 74225 6715025   1534529     1371 2023-12-02 02:00                     180
#> 6 74225 6715025   1534529     1371 2023-12-02 05:00                     180
#>      unit  tm3h
#> 1 Celcius -23.0
#> 2 Celcius -18.3
#> 3 Celcius -14.1
#> 4 Celcius -21.5
#> 5 Celcius -16.0
#> 6 Celcius -13.8
gts_dl_coords(coords = coords, env_layer = "tm1h", start_date = "2023-12-01T23", end_date = "2023-12-02T06")
#>        x       y cellindex altitude       date  time time_resolution_minutes
#> 1  58200 6708225   1542878     1294 2023-12-01 23:00                      60
#> 2  58200 6708225   1542878     1294 2023-12-02 00:00                      60
#> 3  58200 6708225   1542878     1294 2023-12-02 01:00                      60
#> 4  58200 6708225   1542878     1294 2023-12-02 02:00                      60
#> 5  58200 6708225   1542878     1294 2023-12-02 03:00                      60
#> 6  58200 6708225   1542878     1294 2023-12-02 04:00                      60
#> 7  58200 6708225   1542878     1294 2023-12-02 05:00                      60
#> 8  58200 6708225   1542878     1294 2023-12-02 06:00                      60
#> 9  74225 6715025   1534529     1371 2023-12-01 23:00                      60
#> 10 74225 6715025   1534529     1371 2023-12-02 00:00                      60
#> 11 74225 6715025   1534529     1371 2023-12-02 01:00                      60
#> 12 74225 6715025   1534529     1371 2023-12-02 02:00                      60
#> 13 74225 6715025   1534529     1371 2023-12-02 03:00                      60
#> 14 74225 6715025   1534529     1371 2023-12-02 04:00                      60
#> 15 74225 6715025   1534529     1371 2023-12-02 05:00                      60
#> 16 74225 6715025   1534529     1371 2023-12-02 06:00                      60
#>       unit  tm1h
#> 1  Celcius -17.8
#> 2  Celcius -16.3
#> 3  Celcius -15.3
#> 4  Celcius -14.8
#> 5  Celcius -15.1
#> 6  Celcius -14.3
#> 7  Celcius -11.9
#> 8  Celcius -13.8
#> 9  Celcius -14.9
#> 10 Celcius -14.9
#> 11 Celcius -14.5
#> 12 Celcius -14.1
#> 13 Celcius -14.5
#> 14 Celcius -14.3
#> 15 Celcius -11.5
#> 16 Celcius -13.3