Download cellwise values for polygon

Downloads data for each grid cell within a polygon (coordinate system: (EPSG: 25833)). Time resolution is given by the choice of environmental variable (env_layer) and start_date/end_date. E.g. "tm3h" for "2023-12-01T06" is three hour data for temperature collected at 06:00 and "tm" for "2023-12-01" is the daily average temperature.

Note: The API has a cap at ~635'000 values in each download and ends with an error for larger queries.

Usage

gts_dl_polygon(
  polygon,
  env_layer,
  start_date,
  end_date,
  return_raw = FALSE,
  verbose = FALSE
)

Arguments

polygon

A polygon as json rings with spatialReference. Use sf_to_json() to convert from sf til json. Should be in coordinate system EPSG:25833.

env_layer

The quoted abbreviation for the environmental layer to download. E.g. Daily precipitaion = "rr", Temperature = "tm", Snow depth = "sd".

start_date

The start date given as: 'YYYY-MM-DD'. If querying three or one hour data, the hour should be given like this: format: 'YYYY-MM-DDT06'

end_date

The end date. See start_date.

return_raw

Set to TRUE to return the raw results from the query (a list) without coercing to a data.frame.

verbose

Set to TRUE to print the query to the console (default FALSE).

Value

A data frame with the GTS data associated with each grid cell for each step in the time resolution, e.g. daily values or values for each hour.

Examples

## Create a sf polygon and convert to json
library(sf)
xx <- c(58200, 74225, 58200, 74226)
yy <- c(6708225, 6708225, 6715025, 6715025)
pol <- st_as_sf(x = data.frame(x = xx, y = yy), coords = c("x", "y"), crs = 25833)
pol <- st_as_sf(st_as_sfc(st_bbox(pol)))
pol <- gts_sf2json(pol)
## Download data
gts_dl_polygon(polygon = pol, env_layer = "tm", start_date = "2023-12-01", end_date = "2023-12-01")
#>     cellindex altitude       date  time time_resolution_minutes    unit    tm
#> 1     1535708     1245 2023-12-01 00:00                    1440 Celcius -12.9
#> 2     1535709     1314 2023-12-01 00:00                    1440 Celcius -13.5
#> 3     1535710     1177 2023-12-01 00:00                    1440 Celcius -12.5
#> 4     1535711     1083 2023-12-01 00:00                    1440 Celcius -11.9
#> 5     1535712     1096 2023-12-01 00:00                    1440 Celcius -12.1
#> 6     1535713     1100 2023-12-01 00:00                    1440 Celcius -12.1
#> 7     1535714     1136 2023-12-01 00:00                    1440 Celcius -12.4
#> 8     1535715      975 2023-12-01 00:00                    1440 Celcius -11.6
#> 9     1535716     1009 2023-12-01 00:00                    1440 Celcius -11.8
#> 10    1535717      961 2023-12-01 00:00                    1440 Celcius -11.6
#> 11    1535718     1013 2023-12-01 00:00                    1440 Celcius -11.9
#> 12    1535719     1198 2023-12-01 00:00                    1440 Celcius -13.2
#> 13    1535720     1224 2023-12-01 00:00                    1440 Celcius -13.8
#> 14    1535721     1273 2023-12-01 00:00                    1440 Celcius -14.3
#> 15    1535722     1357 2023-12-01 00:00                    1440 Celcius -15.0
#> 16    1535723     1441 2023-12-01 00:00                    1440 Celcius -15.8
#> 17    1536903     1259 2023-12-01 00:00                    1440 Celcius -13.0
#> 18    1536904     1241 2023-12-01 00:00                    1440 Celcius -13.0
#> 19    1536905     1121 2023-12-01 00:00                    1440 Celcius -12.2
#> 20    1536906     1045 2023-12-01 00:00                    1440 Celcius -11.8
#> 21    1536907     1041 2023-12-01 00:00                    1440 Celcius -11.8
#> 22    1536908     1048 2023-12-01 00:00                    1440 Celcius -11.9
#> 23    1536909     1009 2023-12-01 00:00                    1440 Celcius -11.8
#> 24    1536910     1032 2023-12-01 00:00                    1440 Celcius -11.9
#> 25    1536911      922 2023-12-01 00:00                    1440 Celcius -11.5
#> 26    1536912     1014 2023-12-01 00:00                    1440 Celcius -11.9
#> 27    1536913     1102 2023-12-01 00:00                    1440 Celcius -12.5
#> 28    1536914     1139 2023-12-01 00:00                    1440 Celcius -12.8
#> 29    1536915     1140 2023-12-01 00:00                    1440 Celcius -13.0
#> 30    1536916     1205 2023-12-01 00:00                    1440 Celcius -13.8
#> 31    1536917     1270 2023-12-01 00:00                    1440 Celcius -14.4
#> 32    1536918     1436 2023-12-01 00:00                    1440 Celcius -15.8
#> 33    1538098     1200 2023-12-01 00:00                    1440 Celcius -12.8
#> 34    1538099     1188 2023-12-01 00:00                    1440 Celcius -12.8
#> 35    1538100     1093 2023-12-01 00:00                    1440 Celcius -12.1
#> 36    1538101     1092 2023-12-01 00:00                    1440 Celcius -12.1
#> 37    1538102     1176 2023-12-01 00:00                    1440 Celcius -12.8
#> 38    1538103     1085 2023-12-01 00:00                    1440 Celcius -12.2
#> 39    1538104     1091 2023-12-01 00:00                    1440 Celcius -12.3
#> 40    1538105     1114 2023-12-01 00:00                    1440 Celcius -12.5
#> 41    1538106     1010 2023-12-01 00:00                    1440 Celcius -12.0
#> 42    1538107      979 2023-12-01 00:00                    1440 Celcius -11.9
#> 43    1538108     1137 2023-12-01 00:00                    1440 Celcius -12.8
#> 44    1538109     1224 2023-12-01 00:00                    1440 Celcius -13.5
#> 45    1538110     1232 2023-12-01 00:00                    1440 Celcius -13.8
#> 46    1538111     1238 2023-12-01 00:00                    1440 Celcius -13.8
#> 47    1538112     1210 2023-12-01 00:00                    1440 Celcius -13.8
#> 48    1538113     1276 2023-12-01 00:00                    1440 Celcius -14.4
#> 49    1539293     1342 2023-12-01 00:00                    1440 Celcius -13.8
#> 50    1539294     1325 2023-12-01 00:00                    1440 Celcius -13.8
#> 51    1539295     1220 2023-12-01 00:00                    1440 Celcius -13.0
#> 52    1539296     1142 2023-12-01 00:00                    1440 Celcius -12.5
#> 53    1539297     1182 2023-12-01 00:00                    1440 Celcius -12.9
#> 54    1539298     1140 2023-12-01 00:00                    1440 Celcius -12.6
#> 55    1539299     1144 2023-12-01 00:00                    1440 Celcius -12.8
#> 56    1539300     1161 2023-12-01 00:00                    1440 Celcius -12.9
#> 57    1539301     1007 2023-12-01 00:00                    1440 Celcius -12.0
#> 58    1539302     1025 2023-12-01 00:00                    1440 Celcius -12.2
#> 59    1539303     1123 2023-12-01 00:00                    1440 Celcius -12.8
#> 60    1539304     1210 2023-12-01 00:00                    1440 Celcius -13.5
#> 61    1539305     1268 2023-12-01 00:00                    1440 Celcius -14.1
#> 62    1539306     1310 2023-12-01 00:00                    1440 Celcius -14.5
#> 63    1539307     1244 2023-12-01 00:00                    1440 Celcius -14.0
#> 64    1539308     1310 2023-12-01 00:00                    1440 Celcius -14.6
#> 65    1540488     1266 2023-12-01 00:00                    1440 Celcius -13.3
#> 66    1540489     1352 2023-12-01 00:00                    1440 Celcius -14.0
#> 67    1540490     1214 2023-12-01 00:00                    1440 Celcius -13.1
#> 68    1540491     1089 2023-12-01 00:00                    1440 Celcius -12.3
#> 69    1540492     1105 2023-12-01 00:00                    1440 Celcius -12.4
#> 70    1540493     1099 2023-12-01 00:00                    1440 Celcius -12.4
#> 71    1540494     1142 2023-12-01 00:00                    1440 Celcius -12.8
#> 72    1540495     1181 2023-12-01 00:00                    1440 Celcius -13.1
#> 73    1540496     1006 2023-12-01 00:00                    1440 Celcius -12.1
#> 74    1540497     1021 2023-12-01 00:00                    1440 Celcius -12.2
#> 75    1540498     1142 2023-12-01 00:00                    1440 Celcius -13.0
#> 76    1540499     1188 2023-12-01 00:00                    1440 Celcius -13.4
#> 77    1540500     1208 2023-12-01 00:00                    1440 Celcius -13.6
#> 78    1540501     1228 2023-12-01 00:00                    1440 Celcius -13.9
#> 79    1540502     1283 2023-12-01 00:00                    1440 Celcius -14.4
#> 80    1540503     1459 2023-12-01 00:00                    1440 Celcius -15.8
#> 81    1541683     1242 2023-12-01 00:00                    1440 Celcius -13.2
#> 82    1541684     1152 2023-12-01 00:00                    1440 Celcius -12.6
#> 83    1541685     1136 2023-12-01 00:00                    1440 Celcius -12.6
#> 84    1541686     1191 2023-12-01 00:00                    1440 Celcius -13.0
#> 85    1541687     1045 2023-12-01 00:00                    1440 Celcius -12.1
#> 86    1541688     1063 2023-12-01 00:00                    1440 Celcius -12.3
#> 87    1541689     1075 2023-12-01 00:00                    1440 Celcius -12.3
#> 88    1541690     1144 2023-12-01 00:00                    1440 Celcius -12.8
#> 89    1541691     1097 2023-12-01 00:00                    1440 Celcius -12.5
#> 90    1541692      991 2023-12-01 00:00                    1440 Celcius -11.9
#> 91    1541693     1018 2023-12-01 00:00                    1440 Celcius -12.2
#> 92    1541694     1121 2023-12-01 00:00                    1440 Celcius -12.9
#> 93    1541695     1193 2023-12-01 00:00                    1440 Celcius -13.5
#> 94    1541696     1273 2023-12-01 00:00                    1440 Celcius -14.2
#> 95    1541697     1431 2023-12-01 00:00                    1440 Celcius -15.5
#> 96    1541698     1367 2023-12-01 00:00                    1440 Celcius -15.2
#> 97    1542878     1294 2023-12-01 00:00                    1440 Celcius -13.8
#> 98    1542879     1294 2023-12-01 00:00                    1440 Celcius -13.8
#> 99    1542880     1243 2023-12-01 00:00                    1440 Celcius -13.4
#> 100   1542881     1263 2023-12-01 00:00                    1440 Celcius -13.8
#> 101   1542882     1110 2023-12-01 00:00                    1440 Celcius -12.6
#> 102   1542883     1052 2023-12-01 00:00                    1440 Celcius -12.3
#> 103   1542884     1047 2023-12-01 00:00                    1440 Celcius -12.1
#> 104   1542885     1008 2023-12-01 00:00                    1440 Celcius -11.9
#> 105   1542886      994 2023-12-01 00:00                    1440 Celcius -11.9
#> 106   1542887     1080 2023-12-01 00:00                    1440 Celcius -12.5
#> 107   1542888     1105 2023-12-01 00:00                    1440 Celcius -12.8
#> 108   1542889     1184 2023-12-01 00:00                    1440 Celcius -13.5
#> 109   1542890     1334 2023-12-01 00:00                    1440 Celcius -14.8
#> 110   1542891     1506 2023-12-01 00:00                    1440 Celcius -15.9
#> 111   1542892     1459 2023-12-01 00:00                    1440 Celcius -15.8
#> 112   1542893     1481 2023-12-01 00:00                    1440 Celcius -15.9
gts_dl_polygon(polygon = pol, env_layer = "tm1h", start_date = "2023-12-01T00", end_date = "2023-12-01T01")
#>     cellindex altitude       date  time time_resolution_minutes    unit  tm1h
#> 1     1535708     1245 2023-12-01 00:00                      60 Celcius -23.7
#> 2     1535708     1245 2023-12-01 01:00                      60 Celcius -24.3
#> 3     1535709     1314 2023-12-01 00:00                      60 Celcius -22.9
#> 4     1535709     1314 2023-12-01 01:00                      60 Celcius -24.0
#> 5     1535710     1177 2023-12-01 00:00                      60 Celcius -19.9
#> 6     1535710     1177 2023-12-01 01:00                      60 Celcius -21.9
#> 7     1535711     1083 2023-12-01 00:00                      60 Celcius -16.5
#> 8     1535711     1083 2023-12-01 01:00                      60 Celcius -19.3
#> 9     1535712     1096 2023-12-01 00:00                      60 Celcius -15.4
#> 10    1535712     1096 2023-12-01 01:00                      60 Celcius -19.1
#> 11    1535713     1100 2023-12-01 00:00                      60 Celcius -14.2
#> 12    1535713     1100 2023-12-01 01:00                      60 Celcius -18.7
#> 13    1535714     1136 2023-12-01 00:00                      60 Celcius -14.0
#> 14    1535714     1136 2023-12-01 01:00                      60 Celcius -18.9
#> 15    1535715      975 2023-12-01 00:00                      60 Celcius -13.3
#> 16    1535715      975 2023-12-01 01:00                      60 Celcius -18.1
#> 17    1535716     1009 2023-12-01 00:00                      60 Celcius -12.9
#> 18    1535716     1009 2023-12-01 01:00                      60 Celcius -17.8
#> 19    1535717      961 2023-12-01 00:00                      60 Celcius -12.2
#> 20    1535717      961 2023-12-01 01:00                      60 Celcius -16.8
#> 21    1535718     1013 2023-12-01 00:00                      60 Celcius -12.3
#> 22    1535718     1013 2023-12-01 01:00                      60 Celcius -16.2
#> 23    1535719     1198 2023-12-01 00:00                      60 Celcius -13.2
#> 24    1535719     1198 2023-12-01 01:00                      60 Celcius -16.8
#> 25    1535720     1224 2023-12-01 00:00                      60 Celcius -13.2
#> 26    1535720     1224 2023-12-01 01:00                      60 Celcius -17.2
#> 27    1535721     1273 2023-12-01 00:00                      60 Celcius -12.9
#> 28    1535721     1273 2023-12-01 01:00                      60 Celcius -17.1
#> 29    1535722     1357 2023-12-01 00:00                      60 Celcius -15.3
#> 30    1535722     1357 2023-12-01 01:00                      60 Celcius -19.0
#> 31    1535723     1441 2023-12-01 00:00                      60 Celcius -17.9
#> 32    1535723     1441 2023-12-01 01:00                      60 Celcius -20.0
#> 33    1536903     1259 2023-12-01 00:00                      60 Celcius -23.5
#> 34    1536903     1259 2023-12-01 01:00                      60 Celcius -24.7
#> 35    1536904     1241 2023-12-01 00:00                      60 Celcius -22.2
#> 36    1536904     1241 2023-12-01 01:00                      60 Celcius -23.8
#> 37    1536905     1121 2023-12-01 00:00                      60 Celcius -19.0
#> 38    1536905     1121 2023-12-01 01:00                      60 Celcius -21.0
#> 39    1536906     1045 2023-12-01 00:00                      60 Celcius -16.0
#> 40    1536906     1045 2023-12-01 01:00                      60 Celcius -18.3
#> 41    1536907     1041 2023-12-01 00:00                      60 Celcius -15.8
#> 42    1536907     1041 2023-12-01 01:00                      60 Celcius -18.6
#> 43    1536908     1048 2023-12-01 00:00                      60 Celcius -15.3
#> 44    1536908     1048 2023-12-01 01:00                      60 Celcius -18.6
#> 45    1536909     1009 2023-12-01 00:00                      60 Celcius -14.9
#> 46    1536909     1009 2023-12-01 01:00                      60 Celcius -18.4
#> 47    1536910     1032 2023-12-01 00:00                      60 Celcius -14.5
#> 48    1536910     1032 2023-12-01 01:00                      60 Celcius -18.2
#> 49    1536911      922 2023-12-01 00:00                      60 Celcius -13.6
#> 50    1536911      922 2023-12-01 01:00                      60 Celcius -17.5
#> 51    1536912     1014 2023-12-01 00:00                      60 Celcius -13.5
#> 52    1536912     1014 2023-12-01 01:00                      60 Celcius -17.1
#> 53    1536913     1102 2023-12-01 00:00                      60 Celcius -14.8
#> 54    1536913     1102 2023-12-01 01:00                      60 Celcius -17.7
#> 55    1536914     1139 2023-12-01 00:00                      60 Celcius -14.8
#> 56    1536914     1139 2023-12-01 01:00                      60 Celcius -17.9
#> 57    1536915     1140 2023-12-01 00:00                      60 Celcius -13.8
#> 58    1536915     1140 2023-12-01 01:00                      60 Celcius -17.2
#> 59    1536916     1205 2023-12-01 00:00                      60 Celcius -12.9
#> 60    1536916     1205 2023-12-01 01:00                      60 Celcius -16.8
#> 61    1536917     1270 2023-12-01 00:00                      60 Celcius -15.5
#> 62    1536917     1270 2023-12-01 01:00                      60 Celcius -18.5
#> 63    1536918     1436 2023-12-01 00:00                      60 Celcius -18.7
#> 64    1536918     1436 2023-12-01 01:00                      60 Celcius -20.4
#> 65    1538098     1200 2023-12-01 00:00                      60 Celcius -22.5
#> 66    1538098     1200 2023-12-01 01:00                      60 Celcius -23.7
#> 67    1538099     1188 2023-12-01 00:00                      60 Celcius -21.1
#> 68    1538099     1188 2023-12-01 01:00                      60 Celcius -22.7
#> 69    1538100     1093 2023-12-01 00:00                      60 Celcius -18.5
#> 70    1538100     1093 2023-12-01 01:00                      60 Celcius -20.5
#> 71    1538101     1092 2023-12-01 00:00                      60 Celcius -15.9
#> 72    1538101     1092 2023-12-01 01:00                      60 Celcius -18.3
#> 73    1538102     1176 2023-12-01 00:00                      60 Celcius -16.6
#> 74    1538102     1176 2023-12-01 01:00                      60 Celcius -19.5
#> 75    1538103     1085 2023-12-01 00:00                      60 Celcius -16.2
#> 76    1538103     1085 2023-12-01 01:00                      60 Celcius -19.5
#> 77    1538104     1091 2023-12-01 00:00                      60 Celcius -16.0
#> 78    1538104     1091 2023-12-01 01:00                      60 Celcius -19.5
#> 79    1538105     1114 2023-12-01 00:00                      60 Celcius -15.9
#> 80    1538105     1114 2023-12-01 01:00                      60 Celcius -19.3
#> 81    1538106     1010 2023-12-01 00:00                      60 Celcius -15.5
#> 82    1538106     1010 2023-12-01 01:00                      60 Celcius -18.7
#> 83    1538107      979 2023-12-01 00:00                      60 Celcius -14.0
#> 84    1538107      979 2023-12-01 01:00                      60 Celcius -17.2
#> 85    1538108     1137 2023-12-01 00:00                      60 Celcius -15.4
#> 86    1538108     1137 2023-12-01 01:00                      60 Celcius -18.3
#> 87    1538109     1224 2023-12-01 00:00                      60 Celcius -15.8
#> 88    1538109     1224 2023-12-01 01:00                      60 Celcius -18.6
#> 89    1538110     1232 2023-12-01 00:00                      60 Celcius -14.9
#> 90    1538110     1232 2023-12-01 01:00                      60 Celcius -18.1
#> 91    1538111     1238 2023-12-01 00:00                      60 Celcius -13.9
#> 92    1538111     1238 2023-12-01 01:00                      60 Celcius -17.2
#> 93    1538112     1210 2023-12-01 00:00                      60 Celcius -15.8
#> 94    1538112     1210 2023-12-01 01:00                      60 Celcius -18.5
#> 95    1538113     1276 2023-12-01 00:00                      60 Celcius -18.2
#> 96    1538113     1276 2023-12-01 01:00                      60 Celcius -20.0
#> 97    1539293     1342 2023-12-01 00:00                      60 Celcius -22.2
#> 98    1539293     1342 2023-12-01 01:00                      60 Celcius -22.9
#> 99    1539294     1325 2023-12-01 00:00                      60 Celcius -21.3
#> 100   1539294     1325 2023-12-01 01:00                      60 Celcius -22.3
#> 101   1539295     1220 2023-12-01 00:00                      60 Celcius -18.6
#> 102   1539295     1220 2023-12-01 01:00                      60 Celcius -20.6
#> 103   1539296     1142 2023-12-01 00:00                      60 Celcius -16.2
#> 104   1539296     1142 2023-12-01 01:00                      60 Celcius -18.9
#> 105   1539297     1182 2023-12-01 00:00                      60 Celcius -15.8
#> 106   1539297     1182 2023-12-01 01:00                      60 Celcius -19.3
#> 107   1539298     1140 2023-12-01 00:00                      60 Celcius -15.6
#> 108   1539298     1140 2023-12-01 01:00                      60 Celcius -19.9
#> 109   1539299     1144 2023-12-01 00:00                      60 Celcius -15.8
#> 110   1539299     1144 2023-12-01 01:00                      60 Celcius -20.0
#> 111   1539300     1161 2023-12-01 00:00                      60 Celcius -16.3
#> 112   1539300     1161 2023-12-01 01:00                      60 Celcius -19.9
#> 113   1539301     1007 2023-12-01 00:00                      60 Celcius -15.8
#> 114   1539301     1007 2023-12-01 01:00                      60 Celcius -18.7
#> 115   1539302     1025 2023-12-01 00:00                      60 Celcius -14.5
#> 116   1539302     1025 2023-12-01 01:00                      60 Celcius -17.9
#> 117   1539303     1123 2023-12-01 00:00                      60 Celcius -15.8
#> 118   1539303     1123 2023-12-01 01:00                      60 Celcius -18.9
#> 119   1539304     1210 2023-12-01 00:00                      60 Celcius -15.8
#> 120   1539304     1210 2023-12-01 01:00                      60 Celcius -19.0
#> 121   1539305     1268 2023-12-01 00:00                      60 Celcius -15.2
#> 122   1539305     1268 2023-12-01 01:00                      60 Celcius -18.2
#> 123   1539306     1310 2023-12-01 00:00                      60 Celcius -14.2
#> 124   1539306     1310 2023-12-01 01:00                      60 Celcius -17.2
#> 125   1539307     1244 2023-12-01 00:00                      60 Celcius -16.0
#> 126   1539307     1244 2023-12-01 01:00                      60 Celcius -18.7
#> 127   1539308     1310 2023-12-01 00:00                      60 Celcius -18.6
#> 128   1539308     1310 2023-12-01 01:00                      60 Celcius -20.5
#> 129   1540488     1266 2023-12-01 00:00                      60 Celcius -20.5
#> 130   1540488     1266 2023-12-01 01:00                      60 Celcius -21.1
#> 131   1540489     1352 2023-12-01 00:00                      60 Celcius -20.0
#> 132   1540489     1352 2023-12-01 01:00                      60 Celcius -21.0
#> 133   1540490     1214 2023-12-01 00:00                      60 Celcius -18.0
#> 134   1540490     1214 2023-12-01 01:00                      60 Celcius -20.0
#> 135   1540491     1089 2023-12-01 00:00                      60 Celcius -15.4
#> 136   1540491     1089 2023-12-01 01:00                      60 Celcius -18.3
#> 137   1540492     1105 2023-12-01 00:00                      60 Celcius -14.8
#> 138   1540492     1105 2023-12-01 01:00                      60 Celcius -18.5
#> 139   1540493     1099 2023-12-01 00:00                      60 Celcius -14.8
#> 140   1540493     1099 2023-12-01 01:00                      60 Celcius -19.5
#> 141   1540494     1142 2023-12-01 00:00                      60 Celcius -15.3
#> 142   1540494     1142 2023-12-01 01:00                      60 Celcius -20.1
#> 143   1540495     1181 2023-12-01 00:00                      60 Celcius -16.4
#> 144   1540495     1181 2023-12-01 01:00                      60 Celcius -20.1
#> 145   1540496     1006 2023-12-01 00:00                      60 Celcius -16.6
#> 146   1540496     1006 2023-12-01 01:00                      60 Celcius -19.2
#> 147   1540497     1021 2023-12-01 00:00                      60 Celcius -14.9
#> 148   1540497     1021 2023-12-01 01:00                      60 Celcius -18.2
#> 149   1540498     1142 2023-12-01 00:00                      60 Celcius -15.5
#> 150   1540498     1142 2023-12-01 01:00                      60 Celcius -18.9
#> 151   1540499     1188 2023-12-01 00:00                      60 Celcius -15.4
#> 152   1540499     1188 2023-12-01 01:00                      60 Celcius -18.9
#> 153   1540500     1208 2023-12-01 00:00                      60 Celcius -14.6
#> 154   1540500     1208 2023-12-01 01:00                      60 Celcius -17.8
#> 155   1540501     1228 2023-12-01 00:00                      60 Celcius -13.8
#> 156   1540501     1228 2023-12-01 01:00                      60 Celcius -16.8
#> 157   1540502     1283 2023-12-01 00:00                      60 Celcius -16.4
#> 158   1540502     1283 2023-12-01 01:00                      60 Celcius -18.9
#> 159   1540503     1459 2023-12-01 00:00                      60 Celcius -19.5
#> 160   1540503     1459 2023-12-01 01:00                      60 Celcius -20.3
#> 161   1541683     1242 2023-12-01 00:00                      60 Celcius -20.0
#> 162   1541683     1242 2023-12-01 01:00                      60 Celcius -21.1
#> 163   1541684     1152 2023-12-01 00:00                      60 Celcius -19.1
#> 164   1541684     1152 2023-12-01 01:00                      60 Celcius -20.5
#> 165   1541685     1136 2023-12-01 00:00                      60 Celcius -18.3
#> 166   1541685     1136 2023-12-01 01:00                      60 Celcius -20.2
#> 167   1541686     1191 2023-12-01 00:00                      60 Celcius -17.3
#> 168   1541686     1191 2023-12-01 01:00                      60 Celcius -19.9
#> 169   1541687     1045 2023-12-01 00:00                      60 Celcius -15.9
#> 170   1541687     1045 2023-12-01 01:00                      60 Celcius -19.4
#> 171   1541688     1063 2023-12-01 00:00                      60 Celcius -14.9
#> 172   1541688     1063 2023-12-01 01:00                      60 Celcius -19.2
#> 173   1541689     1075 2023-12-01 00:00                      60 Celcius -15.2
#> 174   1541689     1075 2023-12-01 01:00                      60 Celcius -19.5
#> 175   1541690     1144 2023-12-01 00:00                      60 Celcius -16.3
#> 176   1541690     1144 2023-12-01 01:00                      60 Celcius -19.6
#> 177   1541691     1097 2023-12-01 00:00                      60 Celcius -16.6
#> 178   1541691     1097 2023-12-01 01:00                      60 Celcius -18.9
#> 179   1541692      991 2023-12-01 00:00                      60 Celcius -15.2
#> 180   1541692      991 2023-12-01 01:00                      60 Celcius -18.3
#> 181   1541693     1018 2023-12-01 00:00                      60 Celcius -14.8
#> 182   1541693     1018 2023-12-01 01:00                      60 Celcius -18.3
#> 183   1541694     1121 2023-12-01 00:00                      60 Celcius -15.2
#> 184   1541694     1121 2023-12-01 01:00                      60 Celcius -18.6
#> 185   1541695     1193 2023-12-01 00:00                      60 Celcius -15.4
#> 186   1541695     1193 2023-12-01 01:00                      60 Celcius -18.8
#> 187   1541696     1273 2023-12-01 00:00                      60 Celcius -15.8
#> 188   1541696     1273 2023-12-01 01:00                      60 Celcius -18.7
#> 189   1541697     1431 2023-12-01 00:00                      60 Celcius -18.9
#> 190   1541697     1431 2023-12-01 01:00                      60 Celcius -19.5
#> 191   1541698     1367 2023-12-01 00:00                      60 Celcius -20.7
#> 192   1541698     1367 2023-12-01 01:00                      60 Celcius -20.5
#> 193   1542878     1294 2023-12-01 00:00                      60 Celcius -20.4
#> 194   1542878     1294 2023-12-01 01:00                      60 Celcius -21.6
#> 195   1542879     1294 2023-12-01 00:00                      60 Celcius -20.3
#> 196   1542879     1294 2023-12-01 01:00                      60 Celcius -21.6
#> 197   1542880     1243 2023-12-01 00:00                      60 Celcius -20.1
#> 198   1542880     1243 2023-12-01 01:00                      60 Celcius -21.6
#> 199   1542881     1263 2023-12-01 00:00                      60 Celcius -19.8
#> 200   1542881     1263 2023-12-01 01:00                      60 Celcius -21.7
#> 201   1542882     1110 2023-12-01 00:00                      60 Celcius -17.8
#> 202   1542882     1110 2023-12-01 01:00                      60 Celcius -20.9
#> 203   1542883     1052 2023-12-01 00:00                      60 Celcius -15.8
#> 204   1542883     1052 2023-12-01 01:00                      60 Celcius -19.7
#> 205   1542884     1047 2023-12-01 00:00                      60 Celcius -15.6
#> 206   1542884     1047 2023-12-01 01:00                      60 Celcius -19.5
#> 207   1542885     1008 2023-12-01 00:00                      60 Celcius -15.8
#> 208   1542885     1008 2023-12-01 01:00                      60 Celcius -18.8
#> 209   1542886      994 2023-12-01 00:00                      60 Celcius -16.1
#> 210   1542886      994 2023-12-01 01:00                      60 Celcius -18.3
#> 211   1542887     1080 2023-12-01 00:00                      60 Celcius -15.8
#> 212   1542887     1080 2023-12-01 01:00                      60 Celcius -18.5
#> 213   1542888     1105 2023-12-01 00:00                      60 Celcius -16.3
#> 214   1542888     1105 2023-12-01 01:00                      60 Celcius -19.5
#> 215   1542889     1184 2023-12-01 00:00                      60 Celcius -16.2
#> 216   1542889     1184 2023-12-01 01:00                      60 Celcius -19.7
#> 217   1542890     1334 2023-12-01 00:00                      60 Celcius -17.7
#> 218   1542890     1334 2023-12-01 01:00                      60 Celcius -20.0
#> 219   1542891     1506 2023-12-01 00:00                      60 Celcius -18.7
#> 220   1542891     1506 2023-12-01 01:00                      60 Celcius -18.8
#> 221   1542892     1459 2023-12-01 00:00                      60 Celcius -20.9
#> 222   1542892     1459 2023-12-01 01:00                      60 Celcius -20.0
#> 223   1542893     1481 2023-12-01 00:00                      60 Celcius -22.4
#> 224   1542893     1481 2023-12-01 01:00                      60 Celcius -21.2