Download aggregated values for polygon

Downloads and returns data aggregated for the chosen time resolution over grid cells within a polygon (coordinate system: (EPSG: 25833)). Available methods for aggregating are sum, min, max, avg and median. Time resolution is given by the choice of environmental variable (env_layer) and start_date/end_date. E.g. "tm3h" for "2023-12-01T06" is three hour data for temperature collected at 06:00 and "tm" for "2023-12-01" is the daily average temperature. collected at 06:00.

Note: The API has a cap at ~635'000 values in each download and ends with an error for larger queries.

Warning: This function is extremely slow for 1 and 3 hour time steps as the API requires one request per time step when not requesting daily time steps.

Usage

gts_dl_polygon_aggr(
  polygon,
  env_layer,
  start_date,
  end_date,
  method,
  return_raw = FALSE,
  verbose = FALSE
)

Arguments

polygon

A polygon as json rings with spatialReference. Use sf_to_json() to convert from sf til json. Should be in coordinate system EPSG:25833.

env_layer

The quoted abbreviation for the environmental layer to download. E.g. Daily precipitaion = "rr", Temperature = "tm", Snow depth = "sd".

start_date

The start date given as: 'YYYY-MM-DD'. If querying three or one hour data, the hour should be given like this: format: 'YYYY-MM-DDT06'

end_date

The end date. See start_date.

method

The method for aggregating values ("sum", "min", "max", "avg" or "median")

return_raw

Set to TRUE to return the raw results from the query (a list) without coercing to a data.frame (only available for daily time steps).

verbose

Set to TRUE to print the query to the console (default FALSE).

Value

A data frame with the GTS data aggregated across grid cells in the polygon for each step in the time resolution, e.g. daily values or values for each hour.

See also

Examples

## Create a sf polygon and convert to json
library(sf)
xx <- c(58200, 74225, 58200, 74226)
yy <- c(6708225, 6708225, 6715025, 6715025)
pol <- st_as_sf(x = data.frame(x = xx, y = yy), coords = c("x", "y"), crs = 25833)
pol <- st_as_sf(st_as_sfc(st_bbox(pol)))
pol <- gts_sf2json(pol)
## Download data
gts_dl_polygon_aggr(polygon = pol, env_layer = "tm", start_date = "2023-12-01", end_date = "2023-12-04", method = "avg")
#>         date  time time_resolution_minutes    unit    tm_avg
#> 1 2023-12-01 00:00                    1440 Celcius -13.07589
#> 2 2023-12-02 00:00                    1440 Celcius -12.50089
#> 3 2023-12-03 00:00                    1440 Celcius -11.95179
#> 4 2023-12-04 00:00                    1440 Celcius -12.03571
gts_dl_polygon_aggr(polygon = pol, env_layer = "tm1h", start_date = "2023-12-01T06", end_date = "2023-12-01T12", method = "avg")
#>         date  time time_resolution_minutes    unit  tm1h_avg
#> 1 2023-12-01 06:00                      60 Celcius -17.05893
#> 2 2023-12-01 07:00                      60 Celcius -20.72232
#> 3 2023-12-01 08:00                      60 Celcius -20.58750
#> 4 2023-12-01 09:00                      60 Celcius -16.98482
#> 5 2023-12-01 10:00                      60 Celcius -19.91161
#> 6 2023-12-01 11:00                      60 Celcius -14.42768
#> 7 2023-12-01 12:00                      60 Celcius -13.41786
gts_dl_polygon_aggr(polygon = pol, env_layer = "tm3h", start_date = "2023-12-01T06", end_date = "2023-12-01T12", method = "avg")
#>         date  time time_resolution_minutes    unit  tm3h_avg
#> 1 2023-12-01 06:00                     180 Celcius -21.57500
#> 2 2023-12-01 09:00                     180 Celcius -21.32500
#> 3 2023-12-01 12:00                     180 Celcius -17.34554