List parameters available in the GridTimeSeries data

Accesses the available parameters in the GridTimeSeries data and provides a table with the name of parameters, their time resolution, units of measurement and the years of available data.

Usage

gts_parameters()

Value

A table with details the parameters that are available in the GridTimeSeries data. Use the variable names in the table when downloading data.

Examples

## List parameters
gts_parameters()
#>                        Name                                      Fullname
#> 1                        rr                                    Døgnnedbør
#> 2                      rr3d                           Nedbør siste 3 døgn
#> 3                      rr1h                                 Nedbør 1 time
#> 4                      rr3h                                Nedbør 3 timer
#> 5                       rrl                                          Regn
#> 6                     rrl3h                                          Regn
#> 7                 rrprrrxm5                           *Nedbør i % av 5 år
#> 8             rr3hprrr3hxm5 3 timer nedbør i % av 5 års gjentaksnedbør 3t
#> 9                   rrmix3h                              Nysnø og regn 3t
#> 10                     rrsc                                Nedbør som snø
#> 11                   rrsc3h                             Nedbør som snø 3t
#> 12                   darr3h        Disaggregert nedbør seNorge2018_v22.09
#> 13                       tm                                Døgntemperatur
#> 14                     tm1h                             Temperatur 1 time
#> 15                     tm3h                            Temperatur 3 timer
#> 16                   datm3h    Disaggregert temperatur seNorge2018_v22.09
#> 17                      swe                                     Snømengde
#> 18                    swe3h                      Snøens vannekvivalent 3t
#> 19                 snowload                                       Snølast
#> 20                       sd                                      Snødybde
#> 21                     sd3h                              Snødybde 3 timer
#> 22                    swepr                           Snømengde i prosent
#> 23                      age                                  Snøens alder
#> 24                      sca                               Snødekningsgrad
#> 25                      ski                                       Skiføre
#> 26                      qtt                           Regn og snøsmelting
#> 27                    qtt3h             Regn og snøsmelting siste 3 timer
#> 28                    qtt7d                 Regn og snøsmelting siste uke
#> 29                   qtt10d            Regn og snøsmelting siste 10 dager
#> 30                  qttsdai                              Sørpeskredindeks
#> 31                    gwb_q                                     Avrenning
#> 32                  gwb_eva                                   Fordampning
#> 33                  qtt3dls                         *Vanntilførsel 3 døgn
#> 34                     tmgr                             Temperaturendring
#> 35              swechange7d                         Snø endring siste uke
#> 36                      fsw                              Nysnø siste døgn
#> 37                    fsw3h                           Nysnø siste 3 timer
#> 38                    fsw3d                            Nysnø siste 3 døgn
#> 39                    fsw7d                               Nysnø siste uke
#> 40                    sdfsw                                    Nysnødybde
#> 41                  sdfsw3h                      Nysnødybde siste 3 timer
#> 42                      qsw                        Snøsmelting siste døgn
#> 43                    qsw7d                     Snøsmelting sum siste uke
#> 44                    qsw3h                           Snøsmelting 3 timer
#> 45                      lwc                                   Snøtilstand
#> 46                    lwc3h                           Snøtilstand 3 timer
#> 47               gwb_sssdev                          Jordas vannkapasitet
#> 48                  gwb_frd                                       Teledyp
#> 49               gwb_sssrel                            Vannmetning i jord
#> 50                  sdfsw3d                             Nysnødybde 3 døgn
#> 51         gwb_qttprrrxm200                 Vanntilførsel 1 døgn % 200 år
#> 52       gwb_qtt3dprrrxm200                 Vanntilførsel 3 døgn % 200 år
#> 53         gwb_qprqxyrx30yr                     Avrenning i % av maksimum
#> 54     gwb_qttprqttxyrx30yr                 Vanntilførsel 1 døgn i % maks
#> 55 gwb_qtt3dprqtt3dxyrx30yr                 Vanntilførsel 3 døgn i % maks
#> 56      additionalSnowDepth                                 Fokksnøindeks
#> 57          depthHoarIndex1                         Begerkrystallindeks 2
#> 58          depthHoarIndex2                         Begerkrystallindeks 2
#> 59                  gwb_qtt                             HBV Vanntilførsel
#> 60      gwb_landslideindex1                              xJordskredindeks
#> 61      gwb_landslideindex2                             xJordskredindeks2
#> 62                    qttls                                *Vanntilførsel
#> 63               gwb_gwtdev                         Døgnendring Grunnvann
#> 64                  gwb_gwt                                     Grunnvann
#> 65                gwb_qtt3d                      HBV Vanntilførsel 3 døgn
#> 66    gwb_qttprgwb_rryr30yr                          Vanntilførsel 1 døgn
#> 67             gwb_qtt3dlst                          Vanntilførsel 3 døgn
#> 68                  sdfsw7d                            xNysnødybde 7 døgn
#> 69                  swerank                        Snømengde rangert (NY)
#> 70    windDirection10m24h06                          Vindretning 10m døgn
#> 71       windDirection10m1h                        Vindretning 10m 1 time
#> 72       windDirection10m3h                       Vindretning 10m 3 timer
#> 73        windSpeed10m24h06                        Vindhastighet 10m døgn
#> 74           windSpeed10m1h                      Vindhastighet 10m 1 time
#> 75           windSpeed10m3h                     Vindhastighet 10m 3 timer
#> 76                qswenergy           Snøsmelting fra energibalanse model
#> 77              qswenergy3h    Snøsmelting 3timer fra energibalanse model
#> 78           slushflowRatio                            Forholdstall sørpe
#> 79                   fmi10d              Fryse tine grader siste 10 dager
#>    NoDataValue       RawUnit HumanReadableUnit TimeResolutionInMinutes
#> 1        65535            mm                mm                    1440
#> 2        65535            mm                mm                    1440
#> 3        65535            mm                mm                      60
#> 4        65535            mm                mm                     180
#> 5        65535            mm                mm                    1440
#> 6        65535            mm                mm                     180
#> 7        65535             %                 %                    1440
#> 8        65535             %                 %                    1440
#> 9        65535            mm                mm                     180
#> 10       65535            mm                mm                    1440
#> 11       65535            mm                mm                     180
#> 12       65535            mm                mm                     180
#> 13       65535        Kelvin           Celcius                    1440
#> 14       65535        Kelvin           Celcius                      60
#> 15       65535        Kelvin           Celcius                     180
#> 16       65535        Kelvin           Celcius                     180
#> 17       65535            mm                mm                    1440
#> 18       65535            mm                mm                     180
#> 19       65535         kg/m2             kg/m2                    1440
#> 20       65535            mm                cm                    1440
#> 21       65535            mm                cm                     180
#> 22       65535             %                 %                    1440
#> 23         255         Dager             Dager                    1440
#> 24       65535             %                 %                    1440
#> 25         255 Skiføreklasse     Skiføreklasse                    1440
#> 26         255            mm                mm                    1440
#> 27         255            mm                mm                     180
#> 28       65535            mm                mm                    1440
#> 29       65535            mm                mm                    1440
#> 30       65535            mm                mm                    1440
#> 31       65535            mm                mm                    1440
#> 32         255            mm                mm                    1440
#> 33       65535             %                 %                    1440
#> 34       32767    Celcius/10           Celcius                    1440
#> 35       32767            mm                mm                    1440
#> 36         255            mm                mm                    1440
#> 37         255            mm                mm                     180
#> 38       65535            mm                mm                    1440
#> 39       65535            mm                mm                    1440
#> 40       65535            mm                cm                    1440
#> 41       65535            mm                cm                     180
#> 42         255            mm                mm                    1440
#> 43       65535            mm                mm                    1440
#> 44         255            mm                mm                     180
#> 45         255             %                 %                    1440
#> 46         255             %                 %                     180
#> 47       65535            mm                mm                    1440
#> 48       65535            cm                cm                    1440
#> 49       65535             %                 %                    1440
#> 50       65535            mm                cm                    1440
#> 51       65535             %                 %                    1440
#> 52       65535             %                 %                    1440
#> 53       65535             %                 %                    1440
#> 54       65535             %                 %                    1440
#> 55       65535             %                 %                    1440
#> 56       65535            mm                cm                    1440
#> 57       65535     C/m dager         C/m dager                    1440
#> 58       65535         dager             dager                    1440
#> 59       65535            mm                mm                    1440
#> 60       65535         Index             Index                    1440
#> 61       65535         Index             Index                    1440
#> 62       65535             %                 %                    1440
#> 63       32767            mm                mm                    1440
#> 64       65535            mm                mm                    1440
#> 65       65535            mm                mm                    1440
#> 66       65535             %                 %                    1440
#> 67       65535             %                 %                    1440
#> 68       65535            mm                cm                    1440
#> 69         255                                                    1440
#> 70       65535                                                    1440
#> 71       65535                                                      60
#> 72       65535                                                     180
#> 73       65535           m/s               m/s                    1440
#> 74       65535           m/s               m/s                      60
#> 75       65535           m/s               m/s                     180
#> 76       65535            mm                mm                    1440
#> 77       65535            mm                mm                     180
#> 78       65535  Forholdstall      Forholdstall                    1440
#> 79       65535    Døgngrader        Døgngrader                    1440
#>    FirstDateInTimeSerie
#> 1            1957-01-01
#> 2            2010-01-01
#> 3            2010-01-01
#> 4            2010-01-01
#> 5            1957-01-01
#> 6            2010-01-01
#> 7            2018-01-01
#> 8            2010-01-01
#> 9            2010-01-01
#> 10           1957-01-01
#> 11           2010-01-01
#> 12           1971-01-01
#> 13           1957-01-01
#> 14           2010-01-01
#> 15           2010-01-01
#> 16           1971-01-01
#> 17           1957-01-01
#> 18           2010-01-01
#> 19           2000-01-01
#> 20           1957-01-01
#> 21           2010-01-01
#> 22           1957-01-01
#> 23           1957-01-01
#> 24           1957-01-01
#> 25           1957-01-01
#> 26           1957-01-01
#> 27           2010-01-01
#> 28           1957-01-01
#> 29           1957-01-01
#> 30           1957-01-01
#> 31           1957-01-01
#> 32           1957-01-01
#> 33           1957-01-01
#> 34           2018-01-01
#> 35           1957-01-01
#> 36           1957-01-01
#> 37           2010-01-01
#> 38           1957-01-01
#> 39           1957-01-01
#> 40           1957-01-01
#> 41           2010-01-01
#> 42           1957-01-01
#> 43           1957-01-01
#> 44           2010-01-01
#> 45           1957-01-01
#> 46           2010-01-01
#> 47           1957-01-01
#> 48           1957-01-01
#> 49           1957-01-01
#> 50           1957-01-01
#> 51           1957-01-01
#> 52           1957-01-01
#> 53           1957-01-01
#> 54           1957-01-01
#> 55           1957-01-01
#> 56           2021-01-01
#> 57           2010-01-01
#> 58           2010-01-01
#> 59           1957-01-01
#> 60           1957-01-01
#> 61           1957-01-01
#> 62           1957-01-01
#> 63           1957-01-01
#> 64           1957-01-01
#> 65           1957-01-01
#> 66           1957-01-01
#> 67           1957-01-01
#> 68           1957-01-01
#> 69           1957-01-01
#> 70           2010-01-01
#> 71           2010-01-01
#> 72           2010-01-01
#> 73           2010-01-01
#> 74           2010-01-01
#> 75           2010-01-01
#> 76           2019-01-01
#> 77           2019-01-01
#> 78           1957-01-01
#> 79           1957-01-01